requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# LangChain & LangGraph
langchain==0.1.6
langchain-openai==0.0.5
langchain-core==0.1.16
langgraph==0.0.20

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data processing
pandas==2.2.0
numpy==1.26.3

# Environment & Config
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-multipart==0.0.6
